Move login checker to a new file
